refactor!: divide `bool2d` submodule into more files
