+++ src/shapepy/bool2d/shape.py
@@ -11,441 +11,75 @@ from __future__ import annotations
 
 import abc
 from copy import copy
-from fractions import Fraction
-from typing import Any, Iterable, Optional, Tuple, Union
+from typing import Iterable, Optional, Tuple, Union
 
 import numpy as np
-from rbool import Empty
+import rbool
 
 from shapepy.geometry.box import Box
 from shapepy.geometry.jordancurve import JordanCurve
 from shapepy.geometry.point import Point2D
 
 from ..geometry.integral import IntegrateJordan
-from ..geometry.intersection import GeometricIntersectionCurves
 from ..tools import Is, To
+from .base import Empty, SubSetR2, Whole
 
 
-class SuperclassMeta(type):
+class Future:
     """
-    This class is the parent of other classes such
-    still gets the docstrings of their parent
-    """
+    Class that stores methods that are further defined.
+    They are overrided by other methods in __init__.py file
 
-    def __new__(mcs, classname, bases, cls_dict):
-        cls = super().__new__(mcs, classname, bases, cls_dict)
-        for name, member in cls_dict.items():
-            if not getattr(member, "__doc__"):
-                try:
-                    member.__doc__ = getattr(bases[-1], name).__doc__
-                except AttributeError:
-                    pass
-        return cls
+    Although the classes Empty and Whole don't need the
+    child classes to make the union/intersection, or to verify
+    if a SubSetR2 instance is inside Whole for example,
+    the command bellow
+    >>> (0, 0) in Whole()
+    that checks if a point is inside Whole, needs the conversion
+    to a SinglePoint instance, which is not defined in this file
 
+    Another example is the definition of `__add__` method,
+    which must call the function `simplify` after the `__or__`.
+    The function `simplify` must know all the childs classes of SubSetR2,
+    but it would lead to a circular import.
 
-class FollowPath:
-    """
-    Class responsible to compute the final jordan curve
-    result from boolean operation between two simple shapes
-
-    """
-
-    @staticmethod
-    def split_on_intersection(
-        all_group_jordans: Iterable[Iterable[JordanCurve]],
-    ):
-        """
-        Find the intersections between two jordan curves and call split on the
-        nodes which intersects
-        """
-        intersection = GeometricIntersectionCurves([])
-        all_group_jordans = tuple(map(tuple, all_group_jordans))
-        for i, jordansi in enumerate(all_group_jordans):
-            for j in range(i + 1, len(all_group_jordans)):
-                jordansj = all_group_jordans[j]
-                for jordana in jordansi:
-                    for jordanb in jordansj:
-                        intersection |= jordana & jordanb
-        intersection.evaluate()
-        for jordans in all_group_jordans:
-            for jordan in jordans:
-                piece = jordan.piecewise
-                piece.snap(intersection.all_knots[id(jordan)])
-
-    @staticmethod
-    def pursue_path(
-        index_jordan: int, index_segment: int, jordans: Tuple[JordanCurve]
-    ) -> Tuple[Tuple[int]]:
-        """
-        Given a list of jordans, it returns a matrix of integers like
-        [(a1, b1), (a2, b2), (a3, b3), ..., (an, bn)] such
-            End point of jordans[a_{i}].segments[b_{i}]
-            Start point of jordans[a_{i+1}].segments[b_{i+1}]
-        are equal
-
-        The first point (a1, b1) = (index_jordan, index_segment)
-
-        The end point of jordans[an].segments[bn] is equal to
-        the start point of jordans[a1].segments[b1]
-
-        We suppose there's no triple intersection
-        """
-        matrix = []
-        all_segments = [jordan.segments for jordan in jordans]
-        while True:
-            index_segment %= len(all_segments[index_jordan])
-            segment = all_segments[index_jordan][index_segment]
-            if (index_jordan, index_segment) in matrix:
-                break
-            matrix.append((index_jordan, index_segment))
-            last_point = segment.ctrlpoints[-1]
-            possibles = []
-            for i, jordan in enumerate(jordans):
-                if i == index_jordan:
-                    continue
-                if last_point in jordan:
-                    possibles.append(i)
-            if len(possibles) == 0:
-                index_segment += 1
-                continue
-            index_jordan = possibles[0]
-            for j, segj in enumerate(all_segments[index_jordan]):
-                if segj.ctrlpoints[0] == last_point:
-                    index_segment = j
-                    break
-        return tuple(matrix)
-
-    @staticmethod
-    def is_rotation(oneobj: Tuple[Any], other: Tuple[Any]) -> bool:
-        """
-        Tells if a list is equal to another
-        """
-        assert Is.iterable(oneobj)
-        assert Is.iterable(other)
-        oneobj = tuple(oneobj)
-        other = tuple(other)
-        if len(oneobj) != len(other):
-            return False
-        rotation = 0
-        for elem in oneobj:
-            if elem == other[0]:
-                break
-            rotation += 1
-        else:
-            return False
-        nelems = len(other)
-        for i, elem in enumerate(other):
-            j = (i + rotation) % nelems
-            if elem != oneobj[j]:
-                return False
-        return True
-
-    @staticmethod
-    def filter_rotations(matrix: Tuple[Tuple[Any]]):
-        """
-        Remove repeted elements in matrix such they are only rotations
-
-        Example:
-        filter_tuples([[A, B, C], [B, C, A]]) -> [[A, B, C]]
-        filter_tuples([[A, B, C], [C, B, A]]) -> [[A, B, C], [C, B, A]]
-        """
-        filtered = []
-        for line in matrix:
-            for fline in filtered:
-                if FollowPath.is_rotation(line, fline):
-                    break
-            else:
-                filtered.append(line)
-        return tuple(filtered)
-
-    @staticmethod
-    def indexs_to_jordan(
-        jordans: Tuple[JordanCurve], matrix_indexs: Tuple[Tuple[int, int]]
-    ) -> JordanCurve:
-        """
-        Given 'n' jordan curves, and a matrix of integers
-        [(a0, b0), (a1, b1), ..., (am, bm)]
-        Returns a myjordan (JordanCurve instance) such
-        len(myjordan.segments) = matrix_indexs.shape[0]
-        myjordan.segments[i] = jordans[ai].segments[bi]
-        """
-        beziers = []
-        for index_jordan, index_segment in matrix_indexs:
-            new_bezier = jordans[index_jordan].segments[index_segment]
-            new_bezier = copy(new_bezier)
-            beziers.append(new_bezier)
-        new_jordan = JordanCurve(beziers)
-        return new_jordan
-
-    @staticmethod
-    def follow_path(
-        jordans: Tuple[JordanCurve], start_indexs: Tuple[Tuple[int]]
-    ) -> Tuple[JordanCurve]:
-        """
-        Returns a list of jordan curves which is the result
-        of the intersection between 'jordansa' and 'jordansb'
-        """
-        assert all(map(Is.jordan, jordans))
-        bez_indexs = []
-        for ind_jord, ind_seg in start_indexs:
-            indices_matrix = FollowPath.pursue_path(ind_jord, ind_seg, jordans)
-            bez_indexs.append(indices_matrix)
-        bez_indexs = FollowPath.filter_rotations(bez_indexs)
-        new_jordans = []
-        for indices_matrix in bez_indexs:
-            jordan = FollowPath.indexs_to_jordan(jordans, indices_matrix)
-            new_jordans.append(jordan)
-        return tuple(new_jordans)
-
-    @staticmethod
-    def midpoints_one_shape(
-        shapea: BaseShape, shapeb: BaseShape, closed: bool, inside: bool
-    ) -> Tuple[Tuple[int]]:
-        """
-        Returns a matrix [(a0, b0), (a1, b1), ...]
-        such the middle point of
-            shapea.jordans[a0].segments[b0]
-        is inside/outside the shapeb
-
-        If parameter ``closed`` is True, consider a
-        point in boundary is inside.
-        If ``closed=False``, a boundary point is outside
-
-        """
-
-        insiders = []
-        outsiders = []
-        for i, jordan in enumerate(shapea.jordans):
-            for j, segment in enumerate(jordan.segments):
-                mid_point = segment(Fraction(1, 2))
-                if shapeb.contains_point(mid_point, closed):
-                    insiders.append((i, j))
-                else:
-                    outsiders.append((i, j))
-        return tuple(insiders) if inside else tuple(outsiders)
-
-    @staticmethod
-    def midpoints_shapes(
-        shapea: BaseShape, shapeb: BaseShape, closed: bool, inside: bool
-    ) -> Tuple[Tuple[int]]:
-        """
-        This function computes the indexes of the midpoints from
-        both shapes, shifting the indexs of shapeb.jordans
-        """
-        indexsa = FollowPath.midpoints_one_shape(
-            shapea, shapeb, closed, inside
-        )
-        indexsb = FollowPath.midpoints_one_shape(  # pylint: disable=W1114
-            shapeb, shapea, closed, inside
-        )
-        indexsa = list(indexsa)
-        njordansa = len(shapea.jordans)
-        for indjorb, indsegb in indexsb:
-            indexsa.append((njordansa + indjorb, indsegb))
-        return tuple(indexsa)
-
-    @staticmethod
-    def or_shapes(shapea: BaseShape, shapeb: BaseShape) -> Tuple[JordanCurve]:
-        """
-        Computes the set of jordan curves that defines the boundary of
-        the union between the two base shapes
-        """
-        assert Is.instance(shapea, BaseShape)
-        assert Is.instance(shapeb, BaseShape)
-        FollowPath.split_on_intersection([shapea.jordans, shapeb.jordans])
-        indexs = FollowPath.midpoints_shapes(
-            shapea, shapeb, closed=True, inside=False
-        )
-        all_jordans = tuple(shapea.jordans) + tuple(shapeb.jordans)
-        new_jordans = FollowPath.follow_path(all_jordans, indexs)
-        return new_jordans
-
-    @staticmethod
-    def and_shapes(shapea: BaseShape, shapeb: BaseShape) -> Tuple[JordanCurve]:
-        """
-        Computes the set of jordan curves that defines the boundary of
-        the intersection between the two base shapes
-        """
-        assert Is.instance(shapea, BaseShape)
-        assert Is.instance(shapeb, BaseShape)
-        FollowPath.split_on_intersection([shapea.jordans, shapeb.jordans])
-        indexs = FollowPath.midpoints_shapes(
-            shapea, shapeb, closed=False, inside=True
-        )
-        all_jordans = tuple(shapea.jordans) + tuple(shapeb.jordans)
-        new_jordans = FollowPath.follow_path(all_jordans, indexs)
-        return new_jordans
-
-
-# Inherit from object is needed
-# pylint: disable=useless-object-inheritance
-class BaseShape(object, metaclass=SuperclassMeta):
-    """
-    Class which allows operations like:
-     - move
-     - scale
-     - rotation
-     - invert
-     - union
-     - intersection
-     - XOR
+    A solution, which was considered worst is:
+    * Place all the classes and the functions in a single file,
+    so all the classes know all the other classes and we avoid
+    a circular import.
     """
 
-    def __init__(self):
-        pass
+    @staticmethod
+    def unite(subsets: Iterable[SubSetR2]) -> SubSetR2:
+        """
+        Computes the union of some SubSetR2 instances
 
-    @abc.abstractmethod
-    def __invert__(self) -> BaseShape:
-        """Invert shape"""
+        This function is overrided by a function defined
+        in the `shapepy.bool2d.boolean.py` file
+        """
+        raise NotImplementedError
 
-    @abc.abstractmethod
-    def __or__(self, other: BaseShape) -> BaseShape:
-        """Union shapes"""
+    @staticmethod
+    def intersect(subsets: Iterable[SubSetR2]) -> SubSetR2:
+        """
+        Computes the intersection of some SubSetR2 instances
 
-    @abc.abstractmethod
-    def __and__(self, other: BaseShape) -> BaseShape:
-        """Intersection shapes"""
-
-    def __neg__(self) -> BaseShape:
-        """Invert the BaseShape"""
-        return ~self
-
-    def __add__(self, other: BaseShape):
-        """Union of BaseShape"""
-        return self | other
-
-    def __mul__(self, value: BaseShape):
-        """Intersection of BaseShape"""
-        return self & value
-
-    def __sub__(self, value: BaseShape):
-        """Subtraction of BaseShape"""
-        return self & (~value)
-
-    def __xor__(self, other: BaseShape):
-        """XOR of BaseShape"""
-        return (self - other) | (other - self)
-
-    def __bool__(self) -> bool:
-        """Are is positive ?"""
-        return float(self) > 0
-
-
-# pylint: disable=abstract-method
-class SingletonShape(BaseShape):
-    """SingletonShape"""
-
-    __instance = None
-
-    def __new__(cls):
-        if cls.__instance is None:
-            cls.__instance = super(SingletonShape, cls).__new__(cls)
-        return cls.__instance
-
-    def __copy__(self) -> SingletonShape:
-        return self
-
-    def __deepcopy__(self, memo) -> SingletonShape:
-        return self
-
-
-class EmptyShape(SingletonShape):
-    """EmptyShape is a singleton class to represent an empty shape
-
-    Example use
-    -----------
-    >>> from shapepy import EmptyShape
-    >>> empty = EmptyShape()
-    >>> print(empty)
-    EmptyShape
-    >>> print(float(empty))  # Area
-    0.0
-    >>> (0, 0) in empty
-    False
-    """
-
-    def __or__(self, other: BaseShape) -> BaseShape:
-        return copy(other)
-
-    def __and__(self, other: BaseShape) -> BaseShape:
-        return self
-
-    def __sub__(self, other: BaseShape) -> BaseShape:
-        return self
-
-    def __float__(self) -> float:
-        return float(0)
-
-    def __invert__(self) -> BaseShape:
-        return WholeShape()
-
-    def __contains__(self, other: BaseShape) -> bool:
-        return self is other
-
-    def __str__(self) -> str:
-        return "EmptyShape"
-
-    def __repr__(self) -> str:
-        return self.__str__()
-
-
-class WholeShape(SingletonShape):
-    """WholeShape is a singleton class to represent all plane
-
-    Example use
-    -----------
-    >>> from shapepy import WholeShape
-    >>> whole = WholeShape()
-    >>> print(whole)
-    WholeShape
-    >>> print(float(whole))  # Area
-    inf
-    >>> (0, 0) in whole
-    True
-    """
-
-    def __or__(self, other: BaseShape) -> WholeShape:
-        return self
-
-    def __and__(self, other: BaseShape) -> BaseShape:
-        return copy(other)
-
-    def __float__(self) -> float:
-        return float("inf")
-
-    def __invert__(self) -> BaseShape:
-        return EmptyShape()
-
-    def __contains__(self, other: BaseShape) -> bool:
-        return True
-
-    def __str__(self) -> str:
-        return "WholeShape"
-
-    def __repr__(self) -> str:
-        return self.__str__()
-
-    def __sub__(self, other: BaseShape) -> BaseShape:
-        return ~other
+        This function is overrided by a function defined
+        in the `shapepy.bool2d.boolean.py` file
+        """
+        raise NotImplementedError
 
 
 # pylint: disable=no-member
-class DefinedShape(BaseShape):
+class DefinedShape(SubSetR2):
     """
     DefinedShape is the base class for SimpleShape,
     ConnectedShape and DisjointShape
 
     """
 
-    def __init__(self, *args, **kwargs):
-        pass
-
     def __copy__(self) -> DefinedShape:
         return self.__deepcopy__(None)
-
-    def __deepcopy__(self, memo) -> DefinedShape:
-        jordans = tuple(copy(jordan) for jordan in self.jordans)
-        return shape_from_jordans(jordans)
 
     def box(self) -> Box:
         """
@@ -471,44 +105,37 @@ class DefinedShape(BaseShape):
             box |= jordan.box()
         return box
 
-    def __invert__(self) -> BaseShape:
-        return shape_from_jordans(tuple(~jordan for jordan in self.jordans))
-
-    def __or__(self, other: BaseShape) -> BaseShape:
-        assert Is.instance(other, BaseShape)
-        if Is.instance(other, WholeShape):
-            return WholeShape()
-        if Is.instance(other, EmptyShape):
+    def __or__(self, other: SubSetR2) -> SubSetR2:
+        assert Is.instance(other, SubSetR2)
+        if Is.instance(other, Whole):
+            return Whole()
+        if Is.instance(other, Empty):
             return copy(self)
         if other in self:
             return copy(self)
         if self in other:
             return copy(other)
-        new_jordans = FollowPath.or_shapes(self, other)
-        if len(new_jordans) == 0:
-            return WholeShape()
-        return shape_from_jordans(new_jordans)
+        return Future.unite((self, other))
 
-    def __and__(self, other: BaseShape) -> BaseShape:
-        assert Is.instance(other, BaseShape)
-        if Is.instance(other, WholeShape):
+    def __and__(self, other: SubSetR2) -> SubSetR2:
+        assert Is.instance(other, SubSetR2)
+        if Is.instance(other, Whole):
             return copy(self)
-        if Is.instance(other, EmptyShape):
-            return EmptyShape()
+        if Is.instance(other, Empty):
+            return Empty()
         if other in self:
             return copy(other)
         if self in other:
             return copy(self)
-        new_jordans = FollowPath.and_shapes(self, other)
-        if len(new_jordans) == 0:
-            return EmptyShape()
-        return shape_from_jordans(new_jordans)
+        return Future.intersect((self, other))
 
     def __contains__(
-        self, other: Union[Point2D, JordanCurve, BaseShape]
+        self, other: Union[Point2D, JordanCurve, SubSetR2]
     ) -> bool:
-        if Is.instance(other, BaseShape):
+        if Is.instance(other, DefinedShape):
             return self.contains_shape(other)
+        if Is.instance(other, SubSetR2):
+            return Is.instance(other, Empty)
         if Is.instance(other, JordanCurve):
             return self.contains_jordan(other)
         point = To.point(other)
@@ -517,7 +144,7 @@ class DefinedShape(BaseShape):
     def __float__(self) -> float:
         return float(sum(jordan.area for jordan in self.jordans))
 
-    def move(self, point: Point2D) -> BaseShape:
+    def move(self, point: Point2D) -> SubSetR2:
         """
         Moves/translate entire shape by an amount
 
@@ -528,7 +155,7 @@ class DefinedShape(BaseShape):
             The amount to move
 
         :return: The same instance
-        :rtype: BaseShape
+        :rtype: SubSetR2
 
         Example use
         -----------
@@ -542,7 +169,7 @@ class DefinedShape(BaseShape):
             jordan.move(point)
         return self
 
-    def scale(self, xscale: float, yscale: float) -> BaseShape:
+    def scale(self, xscale: float, yscale: float) -> SubSetR2:
         """
         Scales entire shape by an amount
 
@@ -555,7 +182,7 @@ class DefinedShape(BaseShape):
             The amount to scale in vertical direction
 
         :return: The same instance
-        :rtype: BaseShape
+        :rtype: SubSetR2
 
         Example use
         -----------
@@ -568,7 +195,7 @@ class DefinedShape(BaseShape):
             jordan.scale(xscale, yscale)
         return self
 
-    def rotate(self, angle: float, degrees: bool = False) -> BaseShape:
+    def rotate(self, angle: float, degrees: bool = False) -> SubSetR2:
         """
         Rotates entire shape around the origin by an amount
 
@@ -581,7 +208,7 @@ class DefinedShape(BaseShape):
             Flag to indicate if ``angle`` is in radians or degrees
 
         :return: The same instance
-        :rtype: BaseShape
+        :rtype: SubSetR2
 
         Example use
         -----------
@@ -662,7 +289,7 @@ class DefinedShape(BaseShape):
         assert Is.bool(boundary)
         return self._contains_jordan(jordan, boundary)
 
-    def contains_shape(self, other: BaseShape) -> bool:
+    def contains_shape(self, other: SubSetR2) -> bool:
         """
         Checks if the all points of given shape are inside the shape
 
@@ -671,7 +298,7 @@ class DefinedShape(BaseShape):
         Parameters
         ----------
 
-        other : BaseShape
+        other : SubSetR2
             The shape to be verified if is inside
 
         :return: Whether the ``other`` shape is inside or not
@@ -686,11 +313,7 @@ class DefinedShape(BaseShape):
         True
 
         """
-        assert Is.instance(other, BaseShape)
-        if Is.instance(other, EmptyShape):
-            return True
-        if Is.instance(other, WholeShape):
-            return False
+        assert Is.instance(other, DefinedShape)
         return self._contains_shape(other)
 
     @abc.abstractmethod
@@ -704,7 +327,7 @@ class DefinedShape(BaseShape):
         pass
 
     @abc.abstractmethod
-    def _contains_shape(self, other: BaseShape):
+    def _contains_shape(self, other: SubSetR2):
         pass
 
 
@@ -720,8 +343,10 @@ class SimpleShape(DefinedShape):
 
     def __init__(self, jordancurve: JordanCurve):
         assert Is.jordan(jordancurve)
-        super().__init__()
-        self.__set_jordancurve(jordancurve)
+        self.__jordancurve = jordancurve
+
+    def __deepcopy__(self, memo) -> DefinedShape:
+        return SimpleShape(copy(self.__jordancurve))
 
     def __str__(self) -> str:
         area = float(self)
@@ -736,26 +361,26 @@ class SimpleShape(DefinedShape):
         msg = f"Simple shape of area {area:.2f} with {len(vertices)} vertices"
         return msg
 
-    def __eq__(self, other: BaseShape) -> bool:
+    def __eq__(self, other: SubSetR2) -> bool:
         """Compare two shapes
 
         Parameters
         ----------
-        other: BaseShape
+        other: SubSetR2
             The shape to compare
 
-        :raises ValueError: If ``other`` is not a BaseShape instance
+        :raises ValueError: If ``other`` is not a SubSetR2 instance
         """
-        if not Is.instance(other, BaseShape):
+        if not Is.instance(other, SubSetR2):
             raise ValueError
         if not Is.instance(other, SimpleShape):
             return False
         if float(self) != float(other):
             return False
-        return self.jordans[0] == other.jordans[0]
+        return self.__jordancurve == other.jordans[0]
 
     def __invert__(self) -> SimpleShape:
-        return self.__class__(~self.jordans[0])
+        return self.__class__(~self.__jordancurve)
 
     @property
     def jordans(self) -> Tuple[JordanCurve]:
@@ -765,10 +390,6 @@ class SimpleShape(DefinedShape):
         It has only one jordan curve inside it
         """
         return (self.__jordancurve,)
-
-    def __set_jordancurve(self, other: JordanCurve):
-        assert Is.jordan(other)
-        self.__jordancurve = copy(other)
 
     def invert(self) -> SimpleShape:
         """
@@ -819,7 +440,7 @@ class SimpleShape(DefinedShape):
             return False
         inters = piecewise & self.jordans[0]
         inters.evaluate()
-        if inters.all_subsets[id(piecewise)] is not Empty():
+        if inters.all_subsets[id(piecewise)] is not rbool.Empty():
             return True
         knots = sorted(inters.all_knots[id(jordan.piecewise)])
         midknots = ((k0 + k1) / 2 for k0, k1 in zip(knots, knots[1:]))
@@ -880,8 +501,11 @@ class ConnectedShape(DefinedShape):
     """
 
     def __init__(self, subshapes: Tuple[SimpleShape]):
-        super().__init__()
         self.subshapes = subshapes
+
+    def __deepcopy__(self, memo):
+        simples = tuple(map(copy, self.subshapes))
+        return ConnectedShape(simples)
 
     def __float__(self) -> float:
         return sum(map(float, self.subshapes))
@@ -893,8 +517,8 @@ class ConnectedShape(DefinedShape):
     def __repr__(self) -> str:
         return str(self)
 
-    def __eq__(self, other: BaseShape) -> bool:
-        assert Is.instance(other, BaseShape)
+    def __eq__(self, other: SubSetR2) -> bool:
+        assert Is.instance(other, SubSetR2)
         if not Is.instance(other, ConnectedShape):
             return False
         if abs(float(self) - float(other)) > 1e-6:
@@ -991,10 +615,10 @@ class DisjointShape(DefinedShape):
 
     def __new__(cls, subshapes: Tuple[ConnectedShape]):
         subshapes = list(subshapes)
-        while EmptyShape() in subshapes:
-            subshapes.remove(EmptyShape())
+        while Empty() in subshapes:
+            subshapes.remove(Empty())
         if len(subshapes) == 0:
-            return EmptyShape()
+            return Empty()
         for subshape in subshapes:
             assert Is.instance(subshape, (SimpleShape, ConnectedShape))
         if len(subshapes) == 1:
@@ -1004,7 +628,15 @@ class DisjointShape(DefinedShape):
         return instance
 
     def __init__(self, _: Tuple[ConnectedShape]):
-        super().__init__()
+        pass
+
+    def __deepcopy__(self, memo):
+        subshapes = tuple(map(copy, self.subshapes))
+        return DisjointShape(subshapes)
+
+    def __invert__(self):
+        new_jordans = tuple(~jordan for jordan in self.jordans)
+        return shape_from_jordans(new_jordans)
 
     def __float__(self) -> float:
         total = 0
@@ -1012,8 +644,8 @@ class DisjointShape(DefinedShape):
             total += float(subshape)
         return float(total)
 
-    def __eq__(self, other: BaseShape):
-        assert Is.instance(other, BaseShape)
+    def __eq__(self, other: SubSetR2):
+        assert Is.instance(other, SubSetR2)
         if not Is.instance(other, DisjointShape):
             return False
         if float(self) != float(other):
@@ -1113,7 +745,7 @@ class DisjointShape(DefinedShape):
         return self.__subshapes
 
     @subshapes.setter
-    def subshapes(self, values: Tuple[BaseShape]):
+    def subshapes(self, values: Tuple[SubSetR2]):
         for value in values:
             assert Is.instance(value, (SimpleShape, ConnectedShape))
         areas = map(float, values)
@@ -1167,7 +799,7 @@ def divide_connecteds(
     return (connected,) + divide_connecteds(externals)
 
 
-def shape_from_jordans(jordans: Tuple[JordanCurve]) -> BaseShape:
+def shape_from_jordans(jordans: Tuple[JordanCurve]) -> SubSetR2:
     """Returns the correspondent shape
 
     This function don't do entry validation
@@ -1176,7 +808,7 @@ def shape_from_jordans(jordans: Tuple[JordanCurve]) -> BaseShape:
     Example
     ----------
     >>> shape_from_jordans([])
-    EmptyShape
+    Empty
     """
     assert len(jordans) != 0
     simples = tuple(map(SimpleShape, jordans))
