+++ src/shapepy/plot/plot.py
@@ -11,13 +11,8 @@ import matplotlib
 import numpy as np
 from matplotlib import pyplot
 
-from shapepy.bool2d.shape import (
-    BaseShape,
-    ConnectedShape,
-    DisjointShape,
-    EmptyShape,
-    WholeShape,
-)
+from shapepy.bool2d.base import Empty, Whole
+from shapepy.bool2d.shape import ConnectedShape, DisjointShape, SubSetR2
 from shapepy.geometry.jordancurve import JordanCurve
 from shapepy.geometry.segment import Segment
 
@@ -137,19 +132,19 @@ class ShapePloter:
         """
         A wrapper of the matplotlib.pyplot.plt.plot
         """
-        if Is.instance(args[0], BaseShape):
-            return self.plot_shape(args[0], kwargs=kwargs)
+        if Is.instance(args[0], SubSetR2):
+            return self.plot_subset(args[0], kwargs=kwargs)
         return self.gca().plot(*args, **kwargs)
 
     # pylint: disable=too-many-locals
-    def plot_shape(self, shape: BaseShape, *, kwargs):
+    def plot_subset(self, shape: SubSetR2, *, kwargs):
         """
-        Plots a BaseShape, which can be Empty, Whole, Simple, etc
+        Plots a SubSetR2, which can be Empty, Whole, Simple, etc
         """
-        assert Is.instance(shape, BaseShape)
-        if Is.instance(shape, EmptyShape):
+        assert Is.instance(shape, SubSetR2)
+        if Is.instance(shape, Empty):
             return
-        if Is.instance(shape, WholeShape):
+        if Is.instance(shape, Whole):
             self.gca().set_facecolor("#BFFFBF")
             return
         attrs = ["pos_color", "neg_color", "fill_color", "alpha", "marker"]
