+++ tests/bool2d/test_contains.py
@@ -8,13 +8,9 @@ from fractions import Fraction
 import numpy as np
 import pytest
 
+from shapepy.bool2d.base import Empty, Whole
 from shapepy.bool2d.primitive import Primitive
-from shapepy.bool2d.shape import (
-    ConnectedShape,
-    DisjointShape,
-    EmptyShape,
-    WholeShape,
-)
+from shapepy.bool2d.shape import ConnectedShape, DisjointShape
 from shapepy.geometry.factory import FactoryJordan
 
 
@@ -42,10 +38,10 @@ class TestObjectsInEmptyWhole:
     @pytest.mark.order(23)
     @pytest.mark.dependency(depends=["TestObjectsInEmptyWhole::test_begin"])
     def test_singleton(self):
-        empty0 = EmptyShape()
-        whole0 = WholeShape()
-        empty1 = EmptyShape()
-        whole1 = WholeShape()
+        empty0 = Empty()
+        whole0 = Whole()
+        empty1 = Empty()
+        whole1 = Whole()
 
         assert empty0 is empty1
         assert empty0 != whole0
@@ -60,8 +56,8 @@ class TestObjectsInEmptyWhole:
         ]
     )
     def test_empty(self):
-        empty = EmptyShape()
-        whole = WholeShape()
+        empty = Empty()
+        whole = Whole()
         assert empty in empty
         assert empty in whole
 
@@ -73,8 +69,8 @@ class TestObjectsInEmptyWhole:
         ]
     )
     def test_whole(self):
-        empty = EmptyShape()
-        whole = WholeShape()
+        empty = Empty()
+        whole = Whole()
         assert whole not in empty
         assert whole in whole
 
@@ -86,8 +82,8 @@ class TestObjectsInEmptyWhole:
         ]
     )
     def test_point(self):
-        empty = EmptyShape()
-        whole = WholeShape()
+        empty = Empty()
+        whole = Whole()
         for point in [(0, 0), (1, 1), (0, -1)]:
             assert point not in empty
             assert point in whole
@@ -100,8 +96,8 @@ class TestObjectsInEmptyWhole:
         ]
     )
     def test_jordan(self):
-        empty = EmptyShape()
-        whole = WholeShape()
+        empty = Empty()
+        whole = Whole()
 
         vertices = [(0, 0), (1, 0), (0, 1)]
         jordan = FactoryJordan.polygon(vertices)
@@ -121,8 +117,8 @@ class TestObjectsInEmptyWhole:
         ]
     )
     def test_simple_shape(self):
-        empty = EmptyShape()
-        whole = WholeShape()
+        empty = Empty()
+        whole = Whole()
 
         shape = Primitive.triangle()
         assert shape not in empty
@@ -136,8 +132,8 @@ class TestObjectsInEmptyWhole:
         ]
     )
     def test_connected_shape(self):
-        empty = EmptyShape()
-        whole = WholeShape()
+        empty = Empty()
+        whole = Whole()
 
         big_square = Primitive.square(side=2)
         small_square = Primitive.square(side=1)
@@ -153,8 +149,8 @@ class TestObjectsInEmptyWhole:
         ]
     )
     def test_disjoint_shape(self):
-        empty = EmptyShape()
-        whole = WholeShape()
+        empty = Empty()
+        whole = Whole()
 
         square0 = Primitive.square(center=(-3, 0))
         square1 = Primitive.square(center=(3, 0))
@@ -288,7 +284,7 @@ class TestObjectsInSimple:
     @pytest.mark.order(23)
     @pytest.mark.dependency(depends=["TestObjectsInSimple::test_begin"])
     def test_empty(self):
-        empty = EmptyShape()
+        empty = Empty()
         square = Primitive.square()
         assert empty in square
         assert square not in empty
@@ -296,7 +292,7 @@ class TestObjectsInSimple:
     @pytest.mark.order(23)
     @pytest.mark.dependency(depends=["TestObjectsInSimple::test_begin"])
     def test_whole(self):
-        whole = WholeShape()
+        whole = Whole()
         square = Primitive.square()
         assert whole not in square
         assert square in whole
@@ -550,7 +546,7 @@ class TestObjectsInConnected:
     @pytest.mark.order(23)
     @pytest.mark.dependency(depends=["TestObjectsInConnected::test_begin"])
     def test_empty(self):
-        empty = EmptyShape()
+        empty = Empty()
         square = ~Primitive.square()
         assert empty in square
         assert square not in empty
@@ -558,7 +554,7 @@ class TestObjectsInConnected:
     @pytest.mark.order(23)
     @pytest.mark.dependency(depends=["TestObjectsInConnected::test_begin"])
     def test_whole(self):
-        whole = WholeShape()
+        whole = Whole()
         square = ~Primitive.square()
         assert whole not in square
         assert square in whole
@@ -717,7 +713,7 @@ class TestObjectsInDisjoint:
     @pytest.mark.order(23)
     @pytest.mark.dependency(depends=["TestObjectsInDisjoint::test_begin"])
     def test_empty(self):
-        empty = EmptyShape()
+        empty = Empty()
         squarea = Primitive.square(center=(-3, 0))
         squareb = Primitive.square(center=(3, 0))
         disj_shape = DisjointShape([squarea, squareb])
@@ -728,7 +724,7 @@ class TestObjectsInDisjoint:
     @pytest.mark.order(23)
     @pytest.mark.dependency(depends=["TestObjectsInDisjoint::test_begin"])
     def test_whole(self):
-        whole = WholeShape()
+        whole = Whole()
         squarea = Primitive.square(center=(-3, 0))
         squareb = Primitive.square(center=(3, 0))
         disj_shape = DisjointShape([squarea, squareb])
